#61 (#73)

+++ include/Graphics/OpenGL_Bridge/TrianglePrimitive.h
@@ -3,9 +3,9 @@
 
 namespace Barta::Graphics::OpenGL_Bridge {
 struct TrianglePrimitive {
-    GLubyte index1;
-    GLubyte index2;
-    GLubyte index3;
+    GLuint index1;
+    GLuint index2;
+    GLuint index3;
 };
 
 }
